+++ tools/test-js-with-node
@@ -74,6 +74,7 @@ EXEMPT_FILES = make_set(
         "web/src/composebox_typeahead.js",
         "web/src/condense.js",
         "web/src/confirm_dialog.ts",
+        "web/src/copied_tooltip.js",
         "web/src/copy_and_paste.js",
         "web/src/csrf.ts",
         "web/src/css_variables.js",
